Fix bugs found by unit tests

+++ backend/Veterinary.Application/Features/MedicalRecordTextTemplate/Commands/DeleteMedicalRecordTextTemplateCommand.cs
@@ -3,6 +3,8 @@ using MediatR;
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+using Veterinary.Application.Services;
+using Veterinary.Application.Validation.ProblemDetails.Exceptions;
 using Veterinary.Domain.Entities.MedicalRecordEntities;
 
 namespace Veterinary.Application.Features.MedicalRecordTextTemplateFeatures.Commands
@@ -15,14 +17,21 @@ namespace Veterinary.Application.Features.MedicalRecordTextTemplateFeatures.Comm
     public class DeleteMedicalRecordTextTemplateCommandHandler : IRequestHandler<DeleteMedicalRecordTextTemplateCommand, Unit>
     {
         private readonly IMedicalRecordTextTemplateRepository medicalRecordTextTemplateRepository;
+        private readonly IIdentityService identityService;
 
-        public DeleteMedicalRecordTextTemplateCommandHandler(IMedicalRecordTextTemplateRepository medicalRecordTextTemplateRepository)
+        public DeleteMedicalRecordTextTemplateCommandHandler(IMedicalRecordTextTemplateRepository medicalRecordTextTemplateRepository, IIdentityService identityService)
         {
             this.medicalRecordTextTemplateRepository = medicalRecordTextTemplateRepository;
+            this.identityService = identityService;
         }
 
         public async Task<Unit> Handle(DeleteMedicalRecordTextTemplateCommand request, CancellationToken cancellationToken)
         {
+            if (!await identityService.IsInRoleAsync("ManagerDoctor"))
+            {
+                throw new ForbiddenException();
+            }
+
             await medicalRecordTextTemplateRepository.DeleteAsync(request.TemplateId);
             return Unit.Value;
         }
